+++ common/src/comp/controller.rs
@@ -40,6 +40,9 @@ pub enum InventoryAction {
     Use(Slot),
     Sort,
     Collect(Vec3<i32>),
+    // TODO: Not actually inventory-related: refactor to allow sprite interaction without
+    // inventory manipulation!
+    ToggleSpriteLight(VolumePos, bool),
 }
 
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]