+++ server/src/events/mod.rs
@@ -26,7 +26,7 @@ use group_manip::handle_group;
 use information::handle_site_info;
 use interaction::{
     handle_create_sprite, handle_lantern, handle_mine_block, handle_mount, handle_npc_interaction,
-    handle_set_pet_stay, handle_sound, handle_unmount,
+    handle_set_pet_stay, handle_sound, handle_toggle_sprite_light, handle_unmount,
 };
 use inventory_manip::handle_inventory;
 use invite::{handle_invite, handle_invite_response};
@@ -305,6 +305,11 @@ impl Server {
                 ServerEvent::StartTeleporting { entity, portal } => {
                     handle_start_teleporting(self, entity, portal)
                 },
+                ServerEvent::ToggleSpriteLight {
+                    entity,
+                    pos,
+                    enable,
+                } => handle_toggle_sprite_light(self, entity, pos, enable),
             }
         }